+++ src/PSWindowsUpdateGui/Properties/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("PSWindowsUpdateGui.Tests")]
+[assembly: InternalsVisibleTo("PSWindowsUpdateGui.Acceptance")]